build: Upgrade dependencies to latest and adapt to API changes
Bump gpui-component 0.4->0.5.1, reqwest 0.12->0.13.4, rusqlite
0.32->0.40.1, quick-xml 0.36->0.40.1, dirs 5->6 (gpui stays 0.2.2).

Adapt to breaking API changes:
- InputState::multi_line now takes a bool
- rusqlite 0.40 dropped ToSql for usize; bind LIMIT as i64
- scrollable() moved from StyledExt to ScrollableElement::overflow_y_scrollbar

+++ src/db.rs
@@ -96,7 +96,8 @@ impl Database {
              LIMIT ?1",
         )?;
 
-        let items = stmt.query_map([limit], |row| {
+        // rusqlite 0.40 dropped the `ToSql` impl for `usize`; bind as i64.
+        let items = stmt.query_map([limit as i64], |row| {
             let id: i64 = row.get(0)?;
             let timestamp: String = row.get(1)?;
             let method: String = row.get(2)?;
